isaac newton website started
right now im just focusing on funny text. i have a placeholder navigation element, a banner image, and a couple of short descriptions for the website. its gonna be pretty reliant on the writing rather than the design; i dunno how good of an idea that is... i'll make it work.