improved form elements with css, what is it page started (apple joke)
now the elements of the forms are in rows instead of side by side. i tried a couple of different methods for this, but putting each element into div elements worked best and gave me utmost control

for the "what is it" page, i focused on the apple bit. it's actually what made me wanna do a website on isaac newton: i thought itd be funny and charming for an apple to follow you when you scroll and land on newton's head. i got that working, and it works exactly how i imagined it! i might also add some js to make the impact cooler with a graphic. later, ill fill the page up with text and images as well.