better navigation menu and header css
so, im getting started on improving the css. it was a good base, but it needed refining for sure.

first, i started with the naviagtion menu. i moved it out of the header to be its own thing (thats the culprit for the odd-looking edits to the html files). then i used the previous lab as a reference, but i did the styling myself in my own way with a unique hierarchy. it follows you when you scroll, each list element is a button, it shows an underline when you hover, and whichever page you are on has its button darkened (manually by giving that button the "current" id).

then i improved the header. i cropped the newton banner image and made the one for the homepage extra large. i also figured out how to use text shadow from mdn. im really happy with how it looks!

i also made sure each page's title was an h2 element (it was inconsistent before)