created explosion with js (for apple joke)
i remembered in class we made elements hidden or visible depending on a certain condition or something, but i could not find that in any of the lessons nor labs, so i kinda had to relearn this from the ground up in a completely different way. i almost gave up after an hour but then i got it to work!

basically how it works is, the appleJoke.js runs a function every 10 milleseconds that checks if you've scrolled past a certain amount. if you have, it shows the impact image; otherwise it's hidden. this is probably not efficient, but it works, and its cool :D

this may be the last commit if i have no other edits to do.

also:
- adjusted newton's size for the apple joke to work again (some of the css edits i made broke it)